Move viewfinder calculation to ScannerOverlay
